T4.21* blogin poisto ilman tokenia epäonnistuu

+++ controllers/blogs.js
@@ -42,6 +42,12 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
